+++ packages/appcd-plugin/src/tunnel-stream.js
@@ -1,5 +1,5 @@
 import util from 'util';
-
+import { redact } from 'appcd-util';
 import { Writable } from 'stream';
 
 /**
@@ -18,55 +18,27 @@ export default class TunnelStream extends Writable {
 	/**
 	 * Sends a log message to the associated process.
 	 *
-	 * @param {Object} message - The log message.
+	 * @param {Object} message - The data payload containing the log message.
 	 * @param {String} enc - The message encoding. Not used.
 	 * @param {Function} cb - The callback after the write is complete.
 	 * @access public
 	 */
 	_write(message, enc, cb) {
-		if (process.connected && typeof message === 'object') {
-			message.args = [
-				util.format(...message.args.map(a => {
-					return typeof a === 'string' ? a : util.inspect(scrub(a), { colors: true, depth: null });
-				}))
-			];
-
+		if (process.connected && typeof data === 'object') {
 			process.send({
 				type: 'log',
-				message
+				// important! we must clone `message` because redact() will mutate the message
+				// object which may be referenced elsewhere
+				message: {
+					...message,
+					args: [
+						util.format(...message.args.map(a => {
+							return typeof a === 'string' ? a : util.inspect(redact(a), { colors: true, depth: null });
+						}))
+					]
+				}
 			});
 		}
 		cb();
 	}
-}
-
-/**
- * A lookup of various properties that must be redacted during log message serialization.
- * @type {Set}
- */
-const sensitiveProperties = new Set([
-	'clientsecret',
-	'password'
-]);
-
-/**
- * Deeply copies an object and scrubs any potentially sensitive data.
- *
- * @param {Object} src - The source object to copy from.
- * @returns {Object}
- */
-function scrub(src) {
-	if (Array.isArray(src)) {
-		return src.map(scrub);
-	}
-
-	if (src && typeof src === 'object') {
-		const dest = {};
-		for (const [ key, value ] of Object.entries(src)) {
-			dest[key] = sensitiveProperties.has(key.toLowerCase()) ? '<REDACTED>' : scrub(value);
-		}
-		return dest;
-	}
-
-	return src;
 }