+++ packages/appcd-subprocess/CHANGELOG.md
@@ -1,3 +1,7 @@
+# v4.0.1
+
+ * fix: Properly handle IPC send errors when child process goes away.
+
 # v4.0.0 (Jun 12, 2020)
 
  * BREAKING CHANGE: Requires Node.js 10.13.0 or newer.